requirements.txt
gradio
google-generativeai
pytube
PyPDF2
python-dotenv
youtube-transcript-api